+++ src/admin/users.rs
@@ -3,6 +3,7 @@ use std::sync::Arc;
 use axum::extract::{Path, State};
 use axum::http::HeaderMap;
 use axum::Json;
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
@@ -201,8 +202,8 @@ pub(crate) struct CreateKeyReq {
     name: Option<String>,
     enabled: Option<bool>,
     expires_at: Option<String>,
-    #[serde(default)]
-    spend_limit: Option<f64>,
+    #[serde(default, with = "rust_decimal::serde::float_option")]
+    spend_limit: Option<Decimal>,
     #[serde(default)]
     allowed_models: Option<Vec<String>>,
 }
